move fuction to mappings\onnx
