--- tf2onnx/function/__init__.py
@@ -1,23 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT license.
-"""tf2onnx.function module"""
-
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
-from tf2onnx.function.gathernd import gathernd_op
-from tf2onnx.function.lstm_block_cell import lstm_block_cell_op
-from tf2onnx.function.matrixbandpart import matrixbandpart_op
-from tf2onnx.function.range import range_op7
-from tf2onnx.function.select import select_op8
-from tf2onnx.function.sparse_softmax_cross_entropy_with_logits import sparse_softmax_cross_entropy_with_logits_op
-
-__all__ = [
-    "gathernd_op",
-    "lstm_block_cell_op",
-    "matrixbandpart_op",
-    "range_op7",
-    "select_op8",
-    "sparse_softmax_cross_entropy_with_logits_op"
-]
